[Dev] Added an api for the debugging functionality.
The debugging functionality now receives a new debugging class to allow
the integration of a simulator as a library of an external application
(java, python, etc).
Note: this service is supposed to be use through the api, so only when
building the simulator as a library, unlike the inline debugger and the
gdb_sever which are supposed to be used to build standalone programs.

Former-commit-id: bc230cd1b8872353ccc42b52757314ab4c4498cc