[Dev] Added new command line option to set share dir.
Added a new command line option to set the share data directory path. If
the user gives his own share data directory path the system will check
that it is correct and if not correct will try the ancient method
(relative to the binary).

This functionality is necessary for the creation of simulator libraries
which could have a relative path to the share data directory different
from those of traditional binary.


Former-commit-id: 8c49c8c9ac2081a8eb652fc79390cd9c44a389f1