[Dev] Reorganized python armemu library.
The python armemu library has been reorganized, now it clearly contains
two classes 'simulator' and 'variable'.
The paths have been also simplified.
The library now uses the new capability of unisim simulator to set the
path of the share data directory.
This update is under heavy development and is buggy.


Former-commit-id: 38b08c7399841d7b79e51777ed8a489f467c3340