[Dev] Python variables format now depend on simulator ones.
The python variables were previously always managed as strings. Now they
are considered to have the same type than the simulator variable they
are the interface to. When setting a variable the type doesn't really
matter, so the python variable is transformed into a string that will be
used to set the simulator variable.


Former-commit-id: d66e55e18f4d2e425341c8d401b0909f2dc8da1d