[Dev] ARM926EJS-PB: UART now uses reusable sockets.
The idea is that when the simulator stops, halts or aborts the socket
ports it was using were unusable until a certain time (a couple of
minutes). Now they can be reused straight ahead, so the simulator can be
launched again without having to modify its configuration or waiting for
the ports to be released.

Additionally fixed windows compilation.


Former-commit-id: d751563e899425139e7d45b0103ad53f35975536