[Dev] UAPI: created c++ api directory and populated with debug.
A new directory named api is created and programming apis should appear
here. The difference between apis and services is that services are
working inside the simulator and the apis should provide connections to
the external world.

The api directory contains only C++ apis. The C apis are located in the
uapi directory.

The api has been populated with the first api, the debug api.


Former-commit-id: d80fa57457d531137ec30c120af4d86b3274f453