[Dev] ARMEMU: Added method to get the logger output stream as a normal std::ostream.
Thanks to this method libraries now can use (with some restrictions) the
logger facility as a normal std::ostream, so libraries now do not depend
on the kernel logger.
To do so the logger owner must open a debug message before calling
library methods using the logger output stream and close the message
after.

Adapted the dwarf and elf_loader libraries (in unisim/util) to use the
logger output stream instead of the kernel logger, and adapted these
libraries callers (in unisim/service).


Former-commit-id: e5f1e9da0f33f94ce2b6ae0bc389340ec6ac0a20